Added Various things
Added Poetry Type for each crafting workspace item that should be picked up by an NPC.
Added an "AI_NPC2" which will be used to start another behavior tree to make NPCs be different from one another. To accomplish this find the AI Controller Class inside the AI_Character to select which one to use.
In AI_Character "PickUpItem" function, added the logic to set "Is Being Held" to true